+++ src/components/spartan/SInputIncrementBlock/index.ts
@@ -1,0 +1,1 @@
+export { default as SInputIncrementBlock } from './SInputIncrementBlock.vue';