+++ src/components/spartan/index.ts
@@ -36,3 +36,6 @@ export * from './SSidebar';
 export * from './STab';
 export * from './STable';
 export * from './SDataTable';
+export * from './SInputIncrement';
+export * from './SInputIncrementBlock';
+export * from './SSkeleton';