+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `SSkeleton` component implementation.
+
+## 2.0.0-beta.52 - 2024-01-15
+### Fixed
+- `SInputIncrementBlock` component exportation.
+- buttons on `SInputIncrement` avoid submit form.
+
+## 2.0.0-beta.51 - 2024-01-15
+### Added
+- `SInputIncrementBlock` component implementation.
+- `SInputIncrement` now supports `error` prop.
+- `SInputIncrement` component implementation.
 
 ## 2.0.0-beta.50 - 2024-01-15